requirements.txt
fitz
pytesseract
pdfplumber
pdfminer.six
pillow
